MpdBuilder Live profile
Add BandwidthEstimator to estimate the required bandwidth for the
segments.
Also includes unit test for the change.

+++ mpd/base/mpd_utils.cc
@@ -6,14 +6,12 @@
 
 #include "mpd/base/mpd_utils.h"
 
-#include <set>
-
 #include "base/logging.h"
 #include "base/strings/string_number_conversions.h"
 #include "mpd/base/content_protection_element.h"
+#include "mpd/base/media_info.pb.h"
 #include "mpd/base/xml/scoped_xml_ptr.h"
 #include "third_party/libxml/src/include/libxml/tree.h"
-
 
 namespace {
 
@@ -83,7 +81,7 @@ std::string GetCodecs(const MediaInfo& media_info) {
   return "";
 }
 
-std::string SecondsToXmlDuration(float seconds) {
+std::string SecondsToXmlDuration(double seconds) {
   return "PT" + base::DoubleToString(seconds) + "S";
 }
